Use compiled test files (for spectests)
