+++ test/spec/general_test.ex
@@ -1,0 +1,6 @@
+defmodule GeneralSpecTest do
+  use ExUnit.Case, async: true
+  require SpecTestUtils
+
+  SpecTestUtils.generate_tests("general")
+end
