+++ test/spec/mainnet_test.ex
@@ -1,0 +1,6 @@
+defmodule MainnetSpecTest do
+  use ExUnit.Case, async: true
+  require SpecTestUtils
+
+  SpecTestUtils.generate_tests("mainnet")
+end
