+++ test/spec/minimal_cappela_test.ex
@@ -1,0 +1,6 @@
+defmodule MinimalCapellaSpecTest do
+  use ExUnit.Case, async: true
+  require SpecTestUtils
+
+  SpecTestUtils.generate_tests("minimal", "capella")
+end
